Add unit tests for security driver and status handling
- Implemented comprehensive unit tests for the security driver in `test_security.cpp`, covering read/write operations, boundary conditions, and error handling.
- Added tests for status messages and CRC calculations in `test_status_and_contract.cpp`, ensuring independence of literals and correctness of expected address calculations.
- Introduced a custom Unity configuration header `unity_config.h` for output handling during tests.
- Created a script `check_no_production_placeholders.py` to enforce code quality by rejecting production code with deferred-implementation markers.

+++ src/platform/esp32/Esp32Transport.cpp
@@ -1029,7 +1029,9 @@ bool Esp32Transport::_writeEightBits(
   allBitsDelivered = false;
   for (int bit = 7; bit >= 0; --bit) {
     bool bitDelivered = false;
-    if (!_writeBit(((value >> static_cast<uint32_t>(bit)) & 0x01u) != 0u,
+    if (!_writeBit(((static_cast<uint32_t>(value) >>
+                     static_cast<uint32_t>(bit)) &
+                    0x01u) != 0u,
                    timing, clock, &bitDelivered)) {
       allBitsDelivered = bit == 0 && bitDelivered;
       return false;
